backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Yuri Kuniyoshi <e235750@ie.u-ryukyu.ac.jp>"]
readme = "README.md"

packages = [
    { include = "app" }
]


[tool.poetry.dependencies]
python = ">=3.13,<3.14" # Python 3.13を指定する場合はこのように範囲で指定します
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.1"
openai = "^1.35.0"
pydantic = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"] # poetry-core のバージョンを広めに設定
build-backend = "poetry.core.masonry.api"